# apps/web/public/openapi.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "Guteneo — API développeurs",
    "version": "0.3.0",
...
  },
  "servers": [
    {
      "url": "https://guteneo.com",
      "description": "Bêta Guteneo ; activation des canaux et clients OAuth à vérifier via les capacités du service."
    }
  ],
...
  "paths": {
...
    "/api/dispatches/{id}": {
      "get": {
        "operationId": "getDispatch",
        "tags": [
          "Envois"
        ],
        "summary": "Consulter état, événements et approbation",
        "description": "Lire avant toute décision de relance. submission_unknown et submitting peuvent déjà correspondre à une communication acceptée par le fournisseur. Ne jamais recréer automatiquement un envoi, changer de clé ou répéter la soumission fournisseur. L’approbation est null si absente ou expirée. quote_expires_at peut être null. Livraison et décompte sont distincts : consulter faxPricing.settlement pour les fax v3. Un statut delivered ne garantit pas encore le débit final ; la réserve reste conservée si l’usage est en attente.",
        "security": [
          {
            "GuteneoOAuth": [
              "dispatches:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ]
      }
    },
    "/api/dispatches/{id}/confirm": {
      "post": {
        "operationId": "confirmDispatch",
        "tags": [
          "Envois"
        ],
        "summary": "Confirmer un envoi déjà approuvé",
        "description": "Administrateur ou superviseur avec droit de validation requis ; le scope dispatches:send seul ne suffit pas. Sans corps. Nécessite une approbation humaine faite dans le navigateur, liée à l’empreinte immutable et encore valide (au plus 15 min, limitée aussi par le devis). Un accord dans une conversation ne suffit pas. L’assistant ne peut jamais appeler /approve pour la produire. La transition prepared→queued, les réservations de quota/crédit et l’outbox sont atomiques. La réponse ne signifie pas livré. Après un timeout HTTP, relire l’envoi ; seule la même confirmation logique peut réutiliser sa clé. Ne jamais relancer une soumission fournisseur à issue inconnue. Les gates de canal, devis et crédit restent applicables.",
        "security": [
          {
            "GuteneoOAuth": [
              "dispatches:send"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
    "/api/dispatches/{id}/cancel": {
      "post": {
        "operationId": "cancelDispatch",
        "tags": [
          "Envois"
        ],
        "summary": "Annuler avant la soumission",
        "description": "Administrateur ou superviseur avec droit de validation requis ; le scope dispatches:send seul ne suffit pas. Sans corps ni clé d’idempotence requise. Autorisé pour prepared ou queued ; un envoi déjà cancelled reste cancelled. Dès que la soumission commence, 409 CANCELLATION_TOO_LATE : aucune annulation fournisseur ni nouveau message n’est déclenché.",
        "security": [
          {
            "GuteneoOAuth": [
              "dispatches:send"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ]
      }
    },
...
    "/api/dispatches/{id}/renew-quote": {
      "post": {
        "operationId": "renewFaxQuote",
        "summary": "Renouveler un devis fax expiré",
        "description": "Recopie le PDF, le numéro E.164, les options, l’expéditeur et le plafond exacts dans un nouveau devis. Réservé à un fax prepared, expiré ou devenu invalide, sans tentative, fournisseur, outbox ni réservation. Annule l’ancien devis et crée le nouveau atomiquement. Le remplacement est toujours individuel, hors de sa campagne d’origine : annoncer ce changement avant renouvellement. Le manifeste, les anciens membres et leurs empreintes restent inchangés ; l’audit conserve originalCampaignId. Une nouvelle approbation est requise ; aucun envoi, accord ou jeton expert hérité. Idempotence serveur stable par ancien dispatchId : un rejeu renvoie le même nouveau devis. Les erreurs de configuration restent bloquantes. Relire le détail avant action : prepared et attempts vide côté REST ; get_dispatch_status fournit attemptCount=0 explicite côté MCP. Un champ absent ne prouve aucune absence de tentative. L’échéance vient de quote_expires_at en REST et quoteExpiresAt en MCP.",
        "security": [
          {
            "GuteneoOAuth": [
              "dispatches:prepare"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nouveau devis ou résultat idempotent, sans approbation ni envoi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispatch"
                }
              }
            }
          },
          "409": {
            "description": "Envoi non renouvelable, devis encore valable ou configuration indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}": {
      "get": {
        "operationId": "getDocument",
        "tags": [
          "Documents"
        ],
        "summary": "Suivre la vérification d’un PDF",
        "description": "Lecture seule du document conservé et de son état analysis. Si processing, Guteneo poursuit la vérification automatiquement ; consulter de nouveau après retryAfterSeconds. Ne pas réimporter le PDF ni déclencher une analyse à chaque lecture. Seul status=ready permet de préparer un envoi.",
        "security": [
          {
            "GuteneoOAuth": [
              "documents:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "GuteneoOAuth": {
        "type": "oauth2",
        "description": "Client OAuth enregistré requis. Code d’autorisation + PKCE S256, audience=https://guteneo.com/mcp. Jeton d’accès uniquement dans Authorization: Bearer. Ne pas utiliser un ID token, une clé fournisseur ou un token dans une URL. Utiliser state et une URI de retour enregistrée exacte ; ne jamais embarquer de client_secret dans un client public. Connexion navigateur Guteneo préalable et adhésion active obligatoires. En bêta, la politique verified_email exige le claim signé https://guteneo.com/verified_account=true ; la MFA n’est pas obligatoire. La politique historique verified_email_and_mfa conserve son exigence MFA pour les administrateurs. Scopes accordés explicitement ; aucun scope ne permet l’approbation humaine.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://pieper.eu.auth0.com/authorize",
            "tokenUrl": "https://pieper.eu.auth0.com/oauth/token",
            "scopes": {
              "documents:read": "Lister et lire les PDF de l’organisation associée.",
              "documents:write": "Déposer, générer et réanalyser les documents.",
              "dispatches:prepare": "Préparer des envois, créer des campagnes et valider un CSV.",
              "dispatches:send": "Confirmer après approbation humaine ou annuler un envoi. Ne permet jamais d’approuver.",
              "dispatches:read": "Lire les envois, campagnes, expéditeurs et consommation."
            }
          }
        },
        "x-pkce-required": true,
        "x-pkce-method": "S256",
        "x-audience": "https://guteneo.com/mcp"
      }
    },
...
    "schemas": {
...
      "DispatchDetail": {
        "type": "object",
        "properties": {
          "dispatch": {
            "$ref": "#/components/schemas/Dispatch"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderEvent"
            }
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attempt"
            }
          },
          "approval": {
            "type": "object",
            "properties": {
              "fingerprint": {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "fingerprint",
              "expires_at"
            ],
            "nullable": true,
            "description": "Approbation non expirée correspondant à l’empreinte courante, sinon null. Ce champ ne permet pas de créer un consentement."
          }
        },
        "required": [
          "dispatch",
          "events",
          "attempts",
          "approval"
        ]
      },
...
    },
    "responses": {
...
      "Error400": {
        "description": "Requête invalide : VALIDATION_ERROR, INVALID_CURSOR, INVALID_IDEMPOTENCY_KEY ou autre validation métier.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "Authentification absente, invalide ou expirée.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Error404": {
        "description": "Ressource absente dans cette organisation.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Identifiant de diagnostic à conserver avec le code d’erreur ; ne jamais joindre le jeton ou le contenu du document.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
...
}